genie_agent/requirements.txt
mlflow
databricks-sdk
databricks-vectorsearch
langchain
databricks-genai-inference
databricks-langchain
unitycatalog-ai[databricks]
unitycatalog-langchain[databricks]